Proposal for more-generic-generic operations
This is a DRAFT idea of how we could improve the generic operations
further and make them applicable to more scenarios. If we like this
direction I can continue to build this out, I did enough in this PR to
communicate the design ideas.

See the PR comments for my open questions and explanations.