Decouple generic PUT/DELETE controllers from rps
This change updates the generic frontend controllers for async PUT and
DELETE to work based on conv.DataModelInterface instead of the
Radius-specific rp.RadiusDataModel. This means that the generic
controllers can now be used in UCP for any of our purposes there. I
previously already made this change for GET and LIST.

There is some minor cleanup needed for existing controllers to work with
the updates which I included in this PR.

The next step here will be to remove further dependencies from the
defaultoperation package. Right now defaultoperation includes the UNION
of all dependencies that we have in Radius. Now that the base layer has
been refactored this can be slimmed down further.