Set location for Link RP service
Link RP has a bug where URLs for async operations are wrong because they
are missing the location and output a blank URL segment. We've seen this
bug in the past and it's NASTY to try and diagnose if you don't already
know the cause.

The root cause is that we forgot to set a field. Lame.

The only async operation in Link RP right now is the DELETE for Mongo.
I'm not 100% certain why we haven't found the bug before now. The
deployment engine definitely chokes on it so when I changed a PUT
operation to async it explodes. This kind of bug won't be caught by
postman testing because we don't usually do polling when we do that.