Add SDK package as home for client code (#4952)
This change introduces a new top-level package 'pkg/sdk' that contains
client code for interacting with the UCP & Radius APIs. The reason for
the new package is that we need to start calling into UCP from inside
our server APIs. We have code today that accesses the database instead
of calling the API (eg: link RP reading the database of core rp). We
also need to call into the deployment engine from the link rp as part of
the recipes work.

The problem is that all of our client code for interacting with the APIs
today is part of pkg/cli. Rather than introducing cli references to our
RP code, I chose to fix the layering. This is a refactor I've been
wanting to do for a while now.

I also used this opportunity to clean up the proliferation of different
options and APIs that have sprung up organically. Lots of our existing
code futzes around with baseURLs and roundtrippers and other low level
details due to the lack of abstraction.

Now there's one abstraction (sdk.Connection), two functions for creating
a connection (one for testing and one for Kubernetes), and one way (per
generation of SDK) to hook up a generated client to a connection.

This change introduces the new functionality, adds tests where they were
missing before and refactors the client code to use the new types. My
next step will be to introduce sdk.Connection on the server side and
wire it up based on our server side configuration.