Enable value-based recipes (#5050)
This change enables value-based recipes. We originally set the goal of
enabling value-based recipes for Redis, but due the nature of the
deployment engine, they are now enabled for any of our links that
support recipes.

The downside is that it's pretty complicated (see new comments in
deployment processor). Since our existing recipe support is based on
resource binding, all of the concepts get mixed together. There's no
easy to way to one or the other.

There's also some significant new infrastructure needed to power this in
the form of:

- A generic Kubernetes handler that supports any Kubernetes type
- A generic AWS handler (same store)
- Functionality for turning a resource ID into a resource identity

There are many impedance mismatches between the "world of UCP" and the
"world of output resource" that we should close down in follow ups.

This PR also incorporates and supersedes #5031 and #5044.