Poll for deployment complete in CLI every 5 sec (#5087)
This change restores the polling frequency for the deployment to 5
seconds inside the CLI. This must have regressed during the recent
client/SDK update. We previously had this set to 5 seconds.

I noticed the regression visually. Deploying a resource that is fast to
complete (like an environment) still takes 30 seconds (default) to
complete. We poll the "operations" at 5 second intervals, so what
happens is the progress indicators all say "complete" but then the CLI
just kinda chills for another 20 seconds.