Initial commit of `to` helper (#5116)
# Description

We have used
[autorest/to](https://github.com/Azure/go-autorest/tree/main/autorest/to)
and
[azcore/to](https://github.com/Azure/azure-sdk-for-go/tree/main/sdk/azcore/to)
in many places. autorest is being deprecated and we need our own to
helpers. This PR is the initial commit of `to` helper package by
leveraging autorest and azcore implementations.

## Issue reference

<!--
We strive to have all PR being opened based on an issue, where the
problem or feature have been discussed prior to implementation.
-->

Fixes: #5113 

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [ ] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [ ] Unit tests passing
* [ ] Extended the documentation / Created issue for it