HOTFIX - unpopulated providers in environment resource (#5189)
# Description

This PR fixes two bugs:

1. The bug neither handles provider pointer types correctly nor treats
invalid provider type as an error. This PR fixes `CreateEnvProviders` to
handle provider pointer types correctly and returns the error if it is
invalid type. This PR includes the missing unit-tests.

2. `rad env create` doesn't handle aws workspace correctly and unit-test
is incorrect. This is unexpected behavior and codes.

## Issue reference

<!--
We strive to have all PR being opened based on an issue, where the
problem or feature have been discussed prior to implementation.
-->

Fixes: #5188

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [ ] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [ ] Unit tests passing
* [ ] Extended the documentation / Created issue for it