Adding Config loader unit tests (#5278)
# Description

Adding unit tests for config loader tests in #5242. Will update based on
changes to that PR and merge into implementation branch. Unit test
changes are located in
[`environment_test.go`](https://github.com/project-radius/radius/blob/c03879381bb0aedc4a4131947ee832df8258dd4d/pkg/recipes/configloader/environment_test.go)
and
[`resoures_test.go`](https://github.com/project-radius/radius/blob/c03879381bb0aedc4a4131947ee832df8258dd4d/pkg/rp/kube/resources_test.go)

## Issue reference

<!--
We strive to have all PR being opened based on an issue, where the
problem or feature have been discussed prior to implementation.
-->

Fixes: https://dev.azure.com/azure-octo/Incubations/_workitems/edit/6076

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [ ] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [ ] Unit tests passing
* [ ] Extended the documentation / Created issue for it