Set polling frequency for deletion to 5 sec
Fixes: #5164

This change sets the polling frequency of asynchronous delete operations
to 5 seconds for UCP/Radius resources. Currently we can take a very long
time to delete things due to the default polling interval of 30 seconds.
Since this traffic is just to the user's Radius install we're not
concerned about overwhelming the server with a 5 second interval.

I reviewed all of the other places in code where we use async operations
as a client. The AWS and Azure resource code paths still use the 30
second interval. I didn't see a good reason to change this right now.
Everywhere else in code we already use a 5 second interval (recipes)
with the same justification.