Config loader changes for refactoring recipe (#5242)
# Description

Adding Environment Configuration Loader. 

- ConfigurationLoader is responsible for getting configurations needed
to deploy recipes. It calls UCP to fetch information on environment and
application and returns the configuration that represents kubernetes
runtime and cloud provider configuration.

## Issue reference

<!--
We strive to have all PR being opened based on an issue, where the
problem or feature have been discussed prior to implementation.
-->

https://github.com/project-radius/radius/issues/5093
## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [x] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [x] Unit tests passing
* [ ] Extended the documentation / Created issue for it

---------

Co-authored-by: Shruthi Kumar <shruthikumar@microsoft.com>