Make Retry-After header configurable
This change makes the value of the Retry-After header configurable for
async operations. The Retry-After header is used by clients (including
rad CLI) to determine the polling frequency of a long-running operation.

The goal of making this configurable to make some of our CLI operations
feel more responsive as part of #5164. Right now deleting an application
takes approximately a minute which is in line with the default
Retry-After value of 60 seconds.

This PR only does the plumbing of making the value configuable. The next
PR will actually adjust the numbers.