Add generic recipe backend controller
This change adds a new backend controller using Go generics that we can use
for all of our recipe types. The new controller implements the general workflow
needed to execute a recipe, and requires each resource type to implement a
"processor" with its unique functionality.

This PR does not include any implementations of processors and so does not actually hook up the new controller yet. This will be done in a follow up. A few more components of the overall design will also come in separate PRs.