Update prompt width to 40 characters # Description
There's a bug in this script that prevents it from actually retrying.
Running `echo` resets the built-in variable `$?`. Oops.

# Description

The prompt width is currently hard-coded to 20

This PR updates it to 40 so we can display GUIDS without clipping

Long term we may want more customization here, but dropping in a strategic fix to get us unblocked on cloud provider experiences

## Issue reference

<!--
We strive to have all PR being opened based on an issue, where the problem or feature have been discussed prior to implementation.
-->

Fixes: #5193

## Checklist

Please make sure you've  completed the relevant tasks for this PR, out of the following list:

* [ ] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [ ] Unit tests passing
* [ ] Extended the documentation / Created issue for it