Add validation helper for resource processors
This change introduces some validation utilities that we will use while authoring resource processors.

The goal of the validator is to turn our processors into declarative-ish code. It implements all of the
semantics we want for processing the results of executing a recipe. See commends on the Validator type
for details.