Add resource client for links/recipes
This change adds a new resource client type for use in our links/recipes. The new client
only supports deletion, and supports Azure, AWS, and Kubernetes resources.

Additionally the client can infer the API version to use. This is important for recipes, because
we handle a lot of different resource types there and want to avoid hardcoded knowledge. Since
this is just being used for deletion, we really don't care about the API version used, we
just want to pick one.

This client also works with the limitations that we have today in UCP:

- Azure resources go directly to ARM without proxying.
- Kubernetes resources go directly to Kubernetes without proxying.
- We have a hardcoded API version for Radius.

This limitations are OK for now because the code won't change when we improve the client.