Enable fallback workspace for last commands (#5421)
# Description

This change enables the fallback workspace for the last two commands
(env create, and deploy/run). For env create, we actally already had the
right behavior. For deploy, we needed to add support for the resource
group as a CLI arg.

## Issue reference
Part of #4401

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

*[x] Code compiles correctly
* [x] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [x] Unit tests passing
* [ ] Extended the documentation / Created issue for it