Add resource client for links/recipes (#5483)
# Description

This change adds a new resource client type for use in our
links/recipes. The new client only supports deletion, and supports
Azure, AWS, and Kubernetes resources.

Additionally the client can infer the API version to use. This is
important for recipes, because we handle a lot of different resource
types there and want to avoid hardcoded knowledge. Since this is just
being used for deletion, we really don't care about the API version
used, we just want to pick one.

This client also works with the limitations that we have today in UCP:

- Azure resources go directly to ARM without proxying.
- Kubernetes resources go directly to Kubernetes without proxying.
- We have a hardcoded API version for Radius.

This limitations are OK for now because the code won't change when we
improve the client.

This is not hooked up to anything yet, breaking up a larger changeset
into smaller, more reviewable chunks.

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [x] Code compiles correctly
* [x] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [x] Unit tests passing
* [ ] Extended the documentation / Created issue for it