Implement redis processor (#5486)
# Description

This change implements the first 'processor' for one of our links using
the new infrastructure.

This redis implementation implements the new semantics that we agreed to
recently:

- Recipes are the default
- Resource binding is removed
- Values set on the resource can override values provided by a recipe

This is not hooked up to anything yet. That will be done next. 

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [X] Code compiles correctly
* [X] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [X] Unit tests passing
* [ ] Extended the documentation / Created issue for it