This change adds updated recipe support for Dapr State Stores
The major change here is to update the API, datamodel, converter, and add a processor for the dapr state store type.

With the new model we have for recipes, the actual Dapr component is created in the recipe. We still create the component in the manual case.

In addition to that, this change adds logging to the test output covering all of the configurable variables we use (things like the magpie image and the repositiory used for recipes). This should make the e2e of working with our integration tests much better for new contributors.