Provide RuntimeConfiguration to processor
This change adds the RuntimeConfiguration to the set of data provided to a processor. This was a gap we identified while working on Dapr state store support. Since the Dapr state store processor needs to create Kubernetes resources, it needs to know the runtime configuration to get the correct namespace.

This change adds the infrastructure needed to get this data, and updates the unit tests of the controller + existing redis processor. The Dapr state store processor will land in a follow up PR, as several changes need to go in first.