Add patch namespace and split test utils (#5506)
## Description

This change adds the server-side "patch namespace" functionality to a
shared utility package and adds tests for it.

Adding tests triggered a need to refactor our `testutil` package. The
problem is that `testutil` references many different abstractions across
Radius (like store and secrets) which causes a package reference cycle.
Our shared Kubernetes utilities are used by the store, and secret
manager.

The fix is to split the Kubernetes testutils functionality into its own
package. 'utils' packages need to be really granular in Go or else you
end up with cycles like this frequently.

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [x] Code compiles correctly
* [x] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [x] Unit tests passing
* [ ] Extended the documentation / Created issue for it