Add ResourceReference datamodel type and validator
This change adds ResourceReference as part of our data model. This is the type that will be used for all of our resources that expose a '.properties.resources' field. This also adds support to the validator to include this pattern.

Since the redis data model was not changed yet, there's a small workaround in place. This will get removed almost immediately, as we're working on the data model changes for redis already.