Update RAD to Radius application in `rad application` CLI command (#5539)
# Description

RAD is not used as a term/noun in the set of Radius resources. Instead,
we use "Radius application". This PR updates the Cobra command for `rad
application` to switch to "Radius application" instead.

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [x] Code compiles correctly
* [x] Adds necessary unit tests for change
* [x] Adds necessary E2E tests for change
* [x] Unit tests passing
* [x] Extended the documentation / Created issue for it