A fresh coat of paint for rad init
This change updates the UI and flow of rad init to be more streamlined and polished. There are many many small changes in this commit, here are the major changes:

- Prompts for text/selection input will hide after a selection has been made
- Many styling and behavior improvements to text/selection controls
- The text control now includes interactive validation, this improves the experience when a validation failure occurs. Previously a user would be sent back to the start for entering an invalid datum.
- Users are now shown a summary to confirm before installation begins (except in dev mode)
- Users will see an interactive progresss display as installation proceeds
- The set of prompts users are shown has been streamlined to remove redundant options (reinstall radius, cloud providers prompted for existing environments).

I also added usage of bubbletea's testing framework for the new summary control. This test framework is still experimental, and I encountered blocking bugs trying to test the interactive progress control. We should revisit this when possible and add more tests.