Add tests for plane-scoped API scenarios
Part of: #5609

This change includes tests for the "plane-scoped" list API scenarios. The existing code in the list shared operation mostly already works for this scenario, but we need to toggle the query to be recursive. We don't do this by default because it would be incorrect and slow.

Also updated some of the cosmosdb storage tests to better cover this scenario. Everything was working as expected, but the tests themselves weren't actually covering this scenario.