Fix handling of --group for rad env create
Fixes: #5678

The issue was the mishandling the `--group` argument. There were two places the group name was used, for display, and to create the workspace. This uses our existing library function to set the scope based on the resource group.

Additionally I cleaned up the CreateEnvironment and CreateUCPResourceGroup functions of the CLI client. These returned an extra boolean value that would NEVER be false for a success case. I found this will reviewing the code and it seemed workwhile to fix.