Simplify and document CLI error handling
This change turns the `FriendlyError` type into an interface and adds convenience functionality for creating user-facing errors. This allows us more flexibility for packages to define their own types of errors that are handled gracefully by the CLI - for example the console interrupt in the prompt package no longer needs special handling.

Additionally I documented our existing practices for user-facing errors in the development documentation and did a consistency pass of ALL of the existing error messages I could find.