Update to pull request template
This change streamlines our pull request template with an eye towards simplifying our release notes process in the future.

There are two big changes compared to today:

- Getting rid of the checklist
- Adding a 'tell us what type of change this is' selector

We're getting rid of the checklist because as reviewers it does not tell us anything. Many of us do not fill it out (myself included) which is a good sign that it's not helpful and adds extra process overhead without a benefit.

The thing replacing is to intended to:

- Clearly document what kind of change is being made and the expectations.
- Add enforcement for linking issues via a bot (will be done next).
- Aid in the generation of release notes via automation.