Use plane-scoped environment list in CLI
Fixes: #5609

This change enables the plane-scoped list APIs for all resource types, adds a functional test, and updates the rad CLI to use this functionality for discovering environments in `rad init`.

When I was testing this out, I found that the OpenAPI validation was blocking these API calls. I tested this before, and I'm not sure how I didn't notice the break when I was enabling the APIs.

Since we don't have good coverage of this scenario I added some of the first functional tests that make direct API calls