Make UCP routing modular
This change makes UCP's routing modular based on the plane. Each plane type (eg: radius, azure, aws) is given its own nested router so that routes cannot overlap. Routing now occurs in two phases:

- First the plane (or non-plane operation) is selected
- Then the request is delegated to the selected plane

The goal is this change is to decouple the 'host' of UCP from the features and dependencies of each plane, and by doing so make each plane more maintainable.