Remove uses of deployment processor
This change removes the unused SQL Server Database controllers (replaced with shared ones) and updates the Extender controllers to avoid using the linkrp DeploymentProcessor. There is no functionality change in this PR just simplification.

Note: the extender type doesn't actually create any resources. This is why it's safe for us to remove the resource deletion logic... it was never executed in the first place.

The next change will update the listSecrets controllers to remove the DeploymentProcessor.