Remove linkrp deployment processor and secretvalueclient
This change removes the linkrp version of the DeploymentProcessor and the SecretValueClient. We no longer need the DeploymentProcessor in the link rp due to the previous cleanup RPs, so we can remove a lot of complexity (finally). We also no longer use the dynamic lookup features of SecretValueClient so we can remove it.