Add address and pathbase properties to controller options
This change takes a step towards consolidating the functionality for controllers in the Radius Repo. Previously we had three different types for controllers and three different options types. So far we we've consolidated down to a single controller interface and two options types.

This change adds the following properties to the armrpc controller options type:

- Address
- PathBase

These properties are heavily used in UCP but not frequently used in the CoreRP or LinkRP controllers.

Additionally, I make renames to be consistent in using `pathBase` over `basePath`.