Use armrpc options and handler definition
This change updates UCP to use the armrpc definition for controller options and for registering operation handlers. This removes some duplication on the UCP side, where we had created alternative versions of the types used in the RPs.

UCP (intentionally) has some special cases to work with because of the proxying it does. These are now treated as special cases and hooked up directly to the router, rather than trying to force them down the code path intended for RPs.