Update old UCP tests
This change updates some of the **oldest** UCP tests to take advantage of improvements we've made to test infrastructure over the past few months. Additionally, I split these tests into more specific packages and files reflecting their area of testing.

Currently there is a single proxy controller implementation for both Azure and Radius scenarios. In my next change the functionality we provide for Azure and Radius will start to diverge. This change prepares the tests for us to start workign on that.