Add 'rad app connections' command
This change adds the 'rad app connections' command that I have been using in external demos for some time. This displays a text-mode version of the application graph including output resources.

Add build infrastructure
We plan to move this functionality to the server when possible. Right now the CLI can query all of the resources and build the graph itself.