Add publishing of terraform recipes for tests
This change addss support for a Terraform registry to our kubernetes clusters in tests. This uses Terraform's ability to download modules over HTTP without involving the full registry protocol. The OSS Terraform registry implementations are not widely used (< 500 stars) and come with required setup like HTTPS certs. This is a simpler alternative.

The server is NGINX serving files from a Kubernetes ConfigMap. It's possible we will outgrow the size of a ConfigMap in the future, and can migrate to a more full-featured server. The nice thing about the current approach is we didn't need to write any new code.

This also adds automation to publish the recipes as part of the makefile, and our PR runs.

I also renamed directories, logging output, and target names to reflect the separation between bicep and terraform for automation purposes.

I also added some 'dummy' Terraform recipes to prove out the automation end to end. These can be removed when we start writing real tests.