Add terraform recipes test infra
This change adds the remaining plumbing to consume our test terraform recipes from a functional test. I have already written a functional test that uses this new functionality but we need to make some additional product code changes to enable it, so I'm not including it in this PR.

The main change in this PR is renaming terminology where appropriate. Many places where we refer to recipes in test infra (env-vars, log statements, etc) don't clarify whether they are for Bicep or Terraform, which will cause confusion. So the main change here is actually renaming for clarity.

The other change I'm making is in the recommended port for the TF module server for local testing. 8080 is used by one of our RPs, so it's not a good choice. It doesn't really matter which port we choose, but we want to make sure it doesn't cause conflicts.