Fix rad debug-logs
Fixes: #5987

This change fixes a regression in `rad debug-logs`. We changed the labels used in our helm chart and forgot to update the CLI code that queries those labels.

I've also added some more diagnostic information for the case where `rad debug-logs` fails in the future. If we hit another case where no pods are found, it will output some suggested next steps for the user. This is just a safeguard in case we have another regression.