Improve failure reporting in CLI
Fixes: #5988

We have logic in the CLI to detect the Radius install and ping it for overall health before beginning other operations. This is helpful because we can report failures gracefully if the endpoint is down or not installed, before moving on to the actual business logic of whatever the CLI is trying to do.

We have a user report of intermittent failures of this check, likely outside of Radius. More information is needed to debug these failures, and so this PR logs significantly more information for troubleshooting. This is our first time seeing this particular failure mode, and so more insight is needed to find the root cause.