Display HTTP response body on health check failure
Fixes: #6006

This change improves the error handling for the CLI and removes the filtering on JSON that was previously applied. My goal in the previous PR was to avoid printing non-textual data to console and possibly corrupt the terminal.

However, for the investigation we're seeing non-JSON data (`text/plain`) which means the previous version of this can't report it. I'm using a bigger hammer here by filtering out ANSI escape sequences. This should allow us to print the message bodies we get back if they are text-based. For non-textual data it won't be useful regardless but this should at least be safe.