Lower polling frequency for async operations
Fixes: #5164

This change updates our async operations for all of our resources to set a lower value.

Based on my investigation of the issue we're experiencing long delete times because the polling frequency of these operations is set to 60 seconds. They actually complete very quickly but the CLI doesn't check the status for 60 seconds.

We're changing this value to 5 seconds for all of our async operations. Based on feedback we have no reason NOT to optimize for speed right now. We can tweak values additionally in the future.