Fix misuse of log verbosity
I found this issue as part of another ongoing RP, and this is the cleanup. I reviewed all of our uses of `logger.V()`.

What I found is that panics in our codebase were not being logged. Any code that was using the pattern of `logger.V(ucplog.Error).Info(...)` was broken, which meant that a lot of really important logging would never happen.

The problem is with the meaning of `logger.V(...)`. It is NOT possible to use `logger.V(...)` to decrease the verbosity of a message, only to increase it. SO when we were doing `logger.V(ucplog.Error)` we were actually turning it into a debug message and it was being filtered out. Verbosity levels in Logr are additive, you can only add to the default level.

With this in mind, I remove the `ucplog.Error` and `ucplog.Warn` levels because they have the opposite of the desired effect. I also want to say that I really dislike this API design. It's overly complex without being flexible. We should migrate to slog soon, as this is very error-prone.