Increase memory limit for applications-rp
We're seeing restarts due to memory usage in long-haul tests. This is due to execution of Terraform recipes, which can cause a spike in memory usage while the recipe is executing.

Based on the data we're seeing an increase of at least 200mb of working set during recipe execution. We don't have metrics in place yet for Terraform so it's not clear how many Terraform processes this represents (maximum of 3 due to concurrency limit of the background worker).

We may need to make future adjustments here as we add more sophisticated tests, and we will iterate based on data from the long haul tests.