Remove ResourceIdentity
This change removes the ResourceIdentity abstraction from Radius. This is an old piece of tech we built to represent resources across many different resource management systems without a uniform way to refer to them. This has been outdated and not aligned with our thinking until we invented UCP.

After this change we use UCP resource IDs everywhere that refer to an external resource.

There are many changes in this PR aimed at simplifying both OutputResources and the things that they refer to. OutputResources have been streamlined to support two usecases:

- Creating new resources (deployment processor)
- Representing a resource that exists already (recipe, or post-deployment-processor).

The `.CreateResource` field represents all of the data used in resource creation and cleanly segments it from the data we save in the database. Going forward we will also remove the deployment processor and fold the resource creation logic into the components that use it. This is a partial step towards that and removes pieces of the data structure that have a small impact.

Additionally there's a lot of cleanup of testing of output resources in this PR. Many many test baselines have been updated in a uniform way. The best way to ensure that all of the old output resource data was updated was to touch all of these files. In doing so I made them consistent to simplify the test authoring. Believe me, it wasn't fun to go update all ~120 converter test files, but it was easier than chasing each test issue separately.