Improvements and tests for UCP proxy
This change refactors and adds tests for the UCP proxy functionality. This is the kernel of our controllers that proxy requests to either the Azure or the Radius plane. I wrote this code a long time ago during the early days of creating UCP, and have gotten better at Go since then.

I'm most of the way through enabling tracked resources and resource group deletion for Radius. This PR is part of a series taht will be split off from that functionality.