Add 'manual' provisioning to containers
This change adds 'manual' provisioning mode to containers, this means that a container can have the setting `resourceProvisioning: 'manual'` which means that Radius won't create any resources. This also includes the ability to specify `resources: ...` to denote the existing resources associated with the container.

This will be used by the new Kubernetes integration that needs to *track* the existance of a container without deploying it.