Make application and environment configurable
This change makes the application and environment of a Recipe or Deployment configurable through the Kubernetes API.

This includes the logic and tests to handle the case where the application or environment change after creation. For this case we need to delete the existing paired resource for two reasons:

- This will result in a different deployment scope (resource group name is computed).
- It is forbidden in our API to change the application or environment of an existing resource.

This PR does not address the case where a Deployment reacts to a Recipe changing its environment or application. That will be handled in a follow-up.