Add functional test for Kubernetes interop
This change adds a new function test for Kubernetes interop.

As part of this PR I added a new functional test run for the Kubernetes interop tests. It didn't feel like these belonged in our of our existing categories, and it's likely we'll add more in the future.

The major change in this PR is to handle synchronous operations correctly. The APIs we call are documented as asychronous, but can complete synchronously, which requires different handling in the code. I was able to address this without significantly complicating the logic.

There were a few other small fixes needed to get the tutorial code passing.