Fix: Folders with spaces cause the installation script to fail
When you try to run the install script on Windows with user
folders that have spaces in the names, the script fails.

The escape characters just weren't enough to solve the problem.
There's no need for them, you can modify the invoke-expression
calls to handle the situation without escaping spaces.