Add Radius Triage Bot
This is a new GH actions workflow that automates our responses to issues. This was pair-programmed by most of the Radius triage team, with the goal and automating some of our common responses, and making sure contributors can find the right information.

This will respond to new issues of the following types:

- Bug
- Feature
- Maintenance

For these categories it will thank the users and give them links to some of our contributing docs.

We also respond to several labels that we already use in the triage process with helpful guidance links.

Additionally we're trying out labels like 'wont-fix' as a **close** action.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>